remove clearance option
which was only used for the foot mortises

I might restore it later, and use it consistently everywhere.  Using it
in just this one place makes the interface excessively confusing.
VCarve appears to handle the important offsetting for all cuts, though
of course this does not apply to laser-cut boxes.